Build130003052018
edit datamange table
edit delete process
add insert process
fix bugs mysql

+++ admin/delete.data.php
@@ -7,11 +7,15 @@ if($_SESSION['UserID'] == "")
 	}
 $conn = mysqli_connect($server,$username,$password,$dbname);
 $id = $_GET["id"];
-$sql = "DELETE FROM data WHERE data_id = '".$id."' ";
-$query = mysqli_query($conn,$sql);
-  	if(mysqli_affected_rows($conn)) {
-  		 echo "ลบข้อมูลสำเร็จ";
-       header('Refresh: 1; URL=datamanage.php');
-  	}
-  	mysqli_close($conn);
+if (isset($id)) {
+	$sql = "DELETE FROM data WHERE data_id = '".$id."' ";
+	$query = mysqli_query($conn,$sql);
+	  	if(mysqli_affected_rows($conn)) {
+	  		 echo "ลบข้อมูลสำเร็จ";
+	       header('Refresh: 1; URL=datamanage.php');
+	  	}
+	  	mysqli_close($conn);
+}else {
+	header("location:datamanage.php");
+}
 ?>
